Move subscribed tests in CalDavContract to CalendarSharingContract
