+++ src/test/java/com/linagora/dav/contracts/cal/CalendarSharingContract.java
@@ -2927,4 +2927,174 @@ public abstract class CalendarSharingContract {
                 assertThat(json).contains(eventUid);
             });
     }
+
+    @Disabled("Wait for a new image")
+    @Test
+    protected void subscribedCalendarShouldExposeCalendarResourceTypeForCalDavDiscovery() {
+        OpenPaasUser bob = extension().newTestUser();
+        OpenPaasUser alice = extension().newTestUser();
+
+        // GIVEN Bob sets his calendar as publicly readable
+        calDavClient.updateCalendarAcl(bob, "{DAV:}read");
+
+        // AND Alice subscribes to Bob's calendar
+        SubscribedCalendarRequest subscribeRequest = SubscribedCalendarRequest.builder()
+            .id(UUID.randomUUID().toString())
+            .sourceUserId(bob.id())
+            .name("Bob's calendar")
+            .color("#FF0000")
+            .build();
+        calDavClient.subscribeToSharedCalendar(alice, subscribeRequest);
+
+        List<JsonNode> subscribedList = calDavClient.findUserSubscribedCalendars(alice).collectList().block();
+        String sharedCalendarPath = subscribedList.get(0).get("_links").get("self").get("href").asText().replace(".json", "");
+
+        // WHEN Alice does PROPFIND on the subscribed calendar
+        DavResponse response = execute(extension().davHttpClient()
+            .headers(alice::impersonatedBasicAuth)
+            .request(HttpMethod.valueOf("PROPFIND"))
+            .uri(sharedCalendarPath)
+            .send(body("""
+                <?xml version="1.0" encoding="utf-8" ?>
+                <D:propfind xmlns:D="DAV:">
+                  <D:prop>
+                    <D:resourcetype/>
+                  </D:prop>
+                </D:propfind>
+                """)));
+
+        // THEN the resourcetype exposes <cal:calendar/>, making the calendar discoverable by any CalDAV client
+        assertThat(response.status()).isEqualTo(207);
+        assertThat(response.body()).contains(sharedCalendarPath);
+        assertThat(response.body()).contains("<d:collection/>", "<cal:calendar/>");
+    }
+
+    @Disabled("Wait for a new image")
+    @Test
+    protected void readOnlySubscribedCalendarShouldOnlyAdvertiseReadPrivileges() {
+        OpenPaasUser bob = extension().newTestUser();
+        OpenPaasUser alice = extension().newTestUser();
+
+        // GIVEN Bob sets his calendar as read-only
+        calDavClient.updateCalendarAcl(bob, "{DAV:}read");
+
+        // AND Alice subscribes as read-only
+        SubscribedCalendarRequest subscribeRequest = SubscribedCalendarRequest.builder()
+            .id(UUID.randomUUID().toString())
+            .sourceUserId(bob.id())
+            .name("Bob's readonly calendar")
+            .color("#FF0000")
+            .build();
+        calDavClient.subscribeToSharedCalendar(alice, subscribeRequest);
+
+        List<JsonNode> subscribedList = calDavClient.findUserSubscribedCalendars(alice).collectList().block();
+        String sharedCalendarPath = subscribedList.get(0).get("_links").get("self").get("href").asText().replace(".json", "");
+
+        // WHEN Alice requests current-user-privilege-set via PROPFIND on the subscribed calendar
+        DavResponse response = execute(extension().davHttpClient()
+            .headers(headers -> alice.impersonatedBasicAuth(headers)
+                .add("Content-Type", "application/xml"))
+            .request(HttpMethod.valueOf("PROPFIND"))
+            .uri(sharedCalendarPath)
+            .send(body("""
+                <?xml version="1.0" encoding="utf-8" ?>
+                <D:propfind xmlns:D="DAV:">
+                  <D:prop>
+                    <D:current-user-privilege-set/>
+                  </D:prop>
+                </D:propfind>
+                """)));
+
+        // THEN only read privileges are advertised, no write
+        assertThat(response.status()).isEqualTo(207);
+        assertThat(response.body()).contains(sharedCalendarPath);
+        assertThat(response.body()).contains("<d:read/>");
+        assertThat(response.body()).doesNotContain("<d:write/>", "<d:write-content/>", "<d:write-properties/>", "<d:all/>");
+    }
+
+    @Test
+    protected void subscribedResourceCalendarShouldOnlyAdvertiseReadPrivileges() {
+        OpenPaasUser bob = extension().newTestUser();
+        OpenPaasUser alice = extension().newTestUser();
+
+        // GIVEN resource
+        OpenPaaSResource resource = extension().getDockerTwakeCalendarSetupSingleton()
+            .getTwakeCalendarProvisioningService()
+            .createResource("projector", "This is a projector", bob)
+            .block();
+
+        // AND Alice subscribes to the resource calendar
+        SubscribedCalendarRequest subscribeRequest = SubscribedCalendarRequest.builder()
+            .id(UUID.randomUUID().toString())
+            .sourceUserId(resource.id())
+            .name("projector resource calendar")
+            .color("#FF0000")
+            .build();
+        calDavClient.subscribeToSharedCalendar(alice, subscribeRequest);
+
+        List<JsonNode> subscribedList = calDavClient.findUserSubscribedCalendars(alice).collectList().block();
+        String sharedCalendarPath = subscribedList.get(0).get("_links").get("self").get("href").asText().replace(".json", "");
+
+        // WHEN Alice requests current-user-privilege-set via PROPFIND on the subscribed calendar
+        DavResponse response = execute(extension().davHttpClient()
+            .headers(headers -> alice.impersonatedBasicAuth(headers)
+                .add("Content-Type", "application/xml"))
+            .request(HttpMethod.valueOf("PROPFIND"))
+            .uri(sharedCalendarPath)
+            .send(body("""
+                <?xml version="1.0" encoding="utf-8" ?>
+                <D:propfind xmlns:D="DAV:">
+                  <D:prop>
+                    <D:current-user-privilege-set/>
+                  </D:prop>
+                </D:propfind>
+                """)));
+
+        // THEN only read privileges are advertised, no write
+        assertThat(response.status()).isEqualTo(207);
+        assertThat(response.body()).contains(sharedCalendarPath);
+        assertThat(response.body()).contains("<d:read/>");
+        assertThat(response.body()).doesNotContain("<d:write/>", "<d:write-content/>", "<d:write-properties/>", "<d:all/>");
+    }
+
+    @Test
+    void readWriteSubscribedCalendarShouldAdvertiseWritePrivileges() {
+        OpenPaasUser bob = extension().newTestUser();
+        OpenPaasUser alice = extension().newTestUser();
+
+        // GIVEN Bob sets his calendar as writable
+        calDavClient.updateCalendarAcl(bob, "{DAV:}write");
+
+        // AND Alice subscribes as read-write
+        SubscribedCalendarRequest subscribeRequest = SubscribedCalendarRequest.builder()
+            .id(UUID.randomUUID().toString())
+            .sourceUserId(bob.id())
+            .name("Bob's writable calendar")
+            .color("#FF0000")
+            .build();
+        calDavClient.subscribeToSharedCalendar(alice, subscribeRequest);
+
+        List<JsonNode> subscribedList = calDavClient.findUserSubscribedCalendars(alice).collectList().block();
+        String sharedCalendarPath = subscribedList.get(0).get("_links").get("self").get("href").asText().replace(".json", "");
+
+        // WHEN Alice requests current-user-privilege-set via PROPFIND on the subscribed calendar
+        DavResponse response = execute(extension().davHttpClient()
+            .headers(headers -> alice.impersonatedBasicAuth(headers)
+                .add("Content-Type", "application/xml"))
+            .request(HttpMethod.valueOf("PROPFIND"))
+            .uri(sharedCalendarPath)
+            .send(body("""
+                <?xml version="1.0" encoding="utf-8" ?>
+                <D:propfind xmlns:D="DAV:">
+                  <D:prop>
+                    <D:current-user-privilege-set/>
+                  </D:prop>
+                </D:propfind>
+                """)));
+
+        // THEN write privileges are advertised
+        assertThat(response.status()).isEqualTo(207);
+        assertThat(response.body()).contains(sharedCalendarPath);
+        assertThat(response.body()).contains("<d:write/>", "<d:read/>");
+    }
 }
